+++ src/js/components/header/header.react.jsx
@@ -3,15 +3,24 @@ import { NavLink } from "react-router-dom";
 import { bindActionCreators } from "redux";
 import DrumMachineActions from "../../actions/root.actions";
 import { DefaultInput } from "../input/input.react.jsx";
-import { DropDownMenu } from "./dropdown.react.jsx";
+import { DropDownMenu } from "../dropdown/dropdown.react.jsx";
+import { getValueFromPath } from "../../library/natives/object";
 
 export let getMenuItemsFromProps = props => {
-  const { dispatch } = props;
+  const { dispatch, auth } = props;
   const trackActions = bindActionCreators(DrumMachineActions.track, dispatch);
+  let userId = getValueFromPath(auth, "user/uid");
   return [{
     name: "Save Track",
     callback: () => trackActions.saveTrack(),
     condition: () => props.auth && props.auth.user && props.track.write
+  }, {
+    name: "New Track",
+    link: "/"
+  }, {
+    name: "My Tracks",
+    link: `/users/${userId}/tracks/`,
+    condition: () => props.auth && props.auth.user
   }, {
     name: "Logout",
     link: "/user/logout",