+++ src/js/library/firebase/db.js
@@ -4,7 +4,7 @@ let getNewTrackKey = (userId) => {
   return init()
     .then(() =>{
       var database = firebase.database();
-      return database.ref("users/" + userId).child("tracks").push().key      
+      return database.ref(`users/${userId}`).child("tracks").push().key      
     });
 };
 
@@ -12,7 +12,7 @@ let saveTrack = (userId, trackId, json) => {
   return init()
     .then(() => {
       var database = firebase.database();
-      return database.ref("users/" + userId + "/tracks/" + trackId)
+      return database.ref(`users/${userId}/tracks/` + trackId)
         .set(json);
     });
 };
@@ -21,10 +21,28 @@ let loadTrack = (userId, trackId) => {
   return init()
     .then(() => {
       var database = firebase.database();
-      return database.ref("users/" + userId + "/tracks/" + trackId)
+      return database.ref(`users/${userId}/tracks/` + trackId)
         .once("value")
         .then(snapshot => snapshot.val());
     });
 };
 
-export { getNewTrackKey, saveTrack, loadTrack };
+let loadUserTracks = userId => {
+  return init()
+    .then(() => {
+      return firebase.database().ref(`users/${userId}/tracks/`).orderByChild("meta/updatedDate").once("value")
+      .then(function(snapshot) {
+        let tracks = [];
+        snapshot.forEach(shot => {
+          tracks = [
+            ...tracks,
+            shot.val()
+          ]
+        });
+        tracks.reverse();
+        return tracks;
+      });
+    });
+};
+
+export { getNewTrackKey, saveTrack, loadTrack, loadUserTracks };