+++ src/js/library/routing/routing.js
@@ -6,6 +6,8 @@ export let matchesNewPath = (oldPath, newPath) => {
 
 export let matchesUserTrackRoute = path => matchPath(path, "/users/:userId/tracks/:trackId");
 
+export let matchesUserTracksRoute = path => !matchesUserTrackRoute(path) ? matchPath(path, "/users/:userId/tracks/") : false;
+
 export let matchesDefaultTrackRoute = path => matchPath(path, "/");
 
 export let matchesNewTrack = (oldId, newId) => newId !== oldId;