+++ src/js/reducers/root.reducer.js
@@ -11,6 +11,7 @@ import notifications from "./notifications.reducer";
 import instruments from "./instruments.reducer";
 import connections from "./connections.reducer";
 import track from "./track.reducer";
+import tracks from "./tracks.reducer";
 import meta from "./meta.reducer";
 
 const rootReducer = combineReducers({
@@ -25,6 +26,7 @@ const rootReducer = combineReducers({
 	instruments,
 	connections,
 	track,
+	tracks,
 	meta,
 	router: routerReducer
 });